finish modularising database queries, relates #146
